+++ src/Kernel.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App;
 
-use App\Shared\Infrastructure\Symfony\DomainConfigLoader;
+use App\Shared\Infrastructure\Symfony\DependencyInjection\DomainConfigLoader;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
