+++ src/Shared/Infrastructure/Symfony/DependencyInjection/DomainConfigLoader.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Shared\Infrastructure\Symfony;
+namespace App\Shared\Infrastructure\Symfony\DependencyInjection;
 
 use Symfony\Component\Config\Loader\LoaderInterface;
