+++ tests/Unit/Shared/Infrastructure/Symfony/DependencyInjection/DomainConfigLoaderTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Unit\Shared\Infrastructure\Symfony\DependencyInjection;
 
-use App\Shared\Infrastructure\Symfony\DomainConfigLoader;
+use App\Shared\Infrastructure\Symfony\DependencyInjection\DomainConfigLoader;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Config\Loader\LoaderInterface;
